Remove @Entity from an abstract model
