+++ packages/server/src/entity/social-account.ts
@@ -1,7 +1,6 @@
-import { Entity, PrimaryColumn, ManyToOne, JoinColumn } from 'typeorm';
+import { Entity, PrimaryColumn, ManyToOne, JoinColumn, Column } from 'typeorm';
 import { Streamer } from './streamer';
 
-@Entity()
 export abstract class SocialAccount {
   @PrimaryColumn('text')
   id: string;
@@ -9,4 +8,22 @@ export abstract class SocialAccount {
   @ManyToOne(() => Streamer, streamer => streamer.socialAccounts)
   @JoinColumn()
   streamer: Streamer;
+}
+
+@Entity()
+export class TwitterAccount extends SocialAccount {
+  @Column('text')
+  screenName: string;
+}
+
+@Entity()
+export class YoutubeChannel extends SocialAccount {
+  @Column('text')
+  channel: string;
+
+  @Column('text')
+  channelName: string;
+
+  @Column('int')
+  creationOrder: number;
 }
