--- packages/server/src/entity/twitter-account.ts
@@ -1,8 +1,0 @@
-import { Entity, Column } from 'typeorm';
-import { SocialAccount } from './social-account';
-
-@Entity()
-export class TwitterAccount extends SocialAccount {
-  @Column('text')
-  screenName: string;
-}
